move business logic to services
